soundglu: Rework CPU interface updates
Stop using the edge of select to gate updates; instead gate writes and
read side effects on ph0_en.

This was causing the sound self-test to fail when it did a 16-bit CMP
against the sound address register because there was only one edge on
select, causing it to read the low byte for both cycles.