sound: Fix IRQ status/acknowledgement
The IRQ status bit in ES5503 register E0 was inverted, but was always read
by the CPU in the "no IRQ" state because the IRQ was being acknowledged
too early as a side effect of writing SOUNDADRL (and thus "reading" the
register). This caused the sound self-test to pass with the inverted IRQ
bit, but IIgs software would throw an "Unclaimed Sound Interrupt" error.

So, invert the IRQ status bit and add a separate host access enable
between soundglu and ES5503 to control read side effects. Because the
ES5503 sits on a dedicated sound bus, it can't use the typical ph0_en
pattern that we use elsewhere for this.