vgc: Adjust vertical timing and VBlank IRQ
Reduce the visible display to 232 lines, with 16-line top/bottom borders,
filling in the remainder of the 262 total lines with vblank.

Ensure somewhat reasonable periods for vertical front porch and vsync.

These timings probably aren't 100% correct, but they seem to work better
than what was there before.

Also, tweak the VBlank IRQ to trigger on display line 192. Per TN 40,
this seems to be based on the legacy Apple II video timings and not the
actual IIgs vertical blanking period.